% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_cbind.R
\name{agg_cbind}
\alias{agg_cbind}
\title{combine target and predictor data}
\usage{
agg_cbind(data = NULL, date_col = c("Year", "Date"),
  scaling = c("Year", "Day"), aggMethod = c("mean", "max"))
}
\arguments{
\item{date_col}{a two character element vector c("Date", "Date"), where the frist element is the date column from  target.
The second element is the date column from  predictor,
The string elements have to be any of names of lubridate accessor function with capitalizing the first letter,
such as "Date", "Year", "Month", "Day", "Hour", "Minute", "Second", "Week", "Quarter", "Semester", "Am", "Pm", etc.}

\item{scaling}{a two character element vector c("Date", "Date"), where the frist element is the destinate temporal scale for target,
the second element is the destinate temporal scale for predictor
The string elements have to be any of names of lubridate accessor function with capitalizing the first letter,
such as "Date", "Year", "Month", "Day", "Hour", "Minute", "Second", "Week", "Quarter", "Semester", "Am", "Pm", etc}

\item{aggMethod}{a two character element vector c("Date", "Date"), where the frist element is the aggregation method for target,
the second element is the aggregation method for predictor.
The string elements have to be those destinaGroup generic methods such as min, max, mean, etc.(see: ??dplyr::summarise)}
}
\description{
combine target and predictor data
}
\examples{
data(weather_tmax_sf);
data(corn_yield_sf);
target_data <- spatio_fuse(target_stN = corn_yield_sf,
                               data_stN = weather_tmax_sf,
                               parm_nm = "tmax",
                               crs = 2163); target_data
target_data <- target_data \%>\%
  nest(data = c(target, predictor)); target_data
data <- target_data[which(target_data$county == "boone"), ]$data[[1]]; data # humboldt for none; allamakee for one; boone for two
agg_cbind(data,
          date_col = c("Year", "Date"),
          scaling = c("Year","Day"),
          aggMethod = c("mean","max"))
}
