% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbind_TargetDatas.R
\name{cbind_TargetDatas}
\alias{cbind_TargetDatas}
\title{bind multiple stN class object}
\usage{
cbind_TargetDatas(target = NULL, join_by = NULL, ...)
}
\description{
bind multiple stN class object
}
\examples{
x <- list(yield_prcp, yield_tmin, yield_tmax)
cbind_TargetDatas(x, target = "yield", join_by = "Year")
}
