% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateNetwor.R
\name{CreateNetwork}
\alias{CreateNetwork}
\title{Function to get the contacts}
\usage{
CreateNetwork(DF, DTh, coords = c("X", "Y"), DateTime = "DateTime", ID = "ID")
}
\arguments{
\item{DF}{A data.frame object with the information of the GPS Locations.}

\item{DTh}{The distance treshold in meters.}

\item{coords}{the names of the columns with the coordinates in the data.frame object.}

\item{DateTime}{The name of the column with the temporal information o fhte GPS record.}

\item{ID}{The name of the column with the unique Identifier for the subjects.}
}
\description{
Function to get the contacts
}
