% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cropping_cbind.R
\name{cropping_cbind}
\alias{cropping_cbind}
\title{assign site to pol by st_intersection()}
\usage{
cropping_cbind(data_stN = NULL, pol = NULL, parm_nm = "parm_nm")
}
\description{
assign site to pol by st_intersection()
}
\examples{
data(weather_tmax_sf)
data(corn_yield_sf)
pol <- corn_yield_sf \%>\%
rename(target = data) \%>\%
nest(data = c(geometry)) \%>\%
rename(pol = data) \%>\%
filter(county == "boone");pol # humboldt for none; allamakee for one; boone for two
pol <- pol$pol[[1]];pol
parm_nm = "parm_nm"
plot(st_geometry(pol))
plot(st_geometry(data_stN), add = TRUE)
cropping_cbind(data_stN = weather_tmax_sf, pol = pol, parm_nm = "tmax")
}
