% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_adj_nbs.R
\name{get_adj_nbs}
\alias{get_adj_nbs}
\title{Function to get the contacts}
\usage{
get_adj_nbs(a, predicate = "st_touches", dist = 10)
}
\arguments{
\item{a}{A sf multipolygon object}

\item{predicate}{the predicate function to do for the proximity, default is st_touches}

\item{dist}{for predicate fnctions based on distance}
}
\description{
Function to get the contacts
}
