% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HexGrid.R
\name{HexGrid}
\alias{HexGrid}
\title{Function to Create a Hexagonal grid in a spatial field}
\usage{
HexGrid(cellsize, Shp)
}
\arguments{
\item{Shp}{A shapefile which will be used as the area for the grids, must be in UTM}

\item{mycellsize}{The size of each cell of the grid.}
}
\description{
Function to Create a Hexagonal grid in a spatial field
}
