% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HexMap.R
\name{HexMap}
\alias{HexMap}
\title{Function to Plot a Hexagonal grid map of counts}
\usage{
HexMap(Hex, Points)
}
\arguments{
\item{Hex}{A hexagon grid, see HexGrid function on how to create it}

\item{Points}{A set of points that will be counted by cell}
}
\description{
Function to Plot a Hexagonal grid map of counts
}
