% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pol_cropping.R
\name{pol_cropping}
\alias{pol_cropping}
\title{assign site to pol by st_intersection()}
\usage{
pol_cropping(data_stN = NULL, pol = NULL, parm_nm = "parm_nm")
}
\description{
assign site to pol by st_intersection()
}
\examples{
data(water_Temperature_sf)
data(corn_yield_sf)
corn_yield_st <- corn_yield_sf
water_Temperature_st <- water_Temperature_sf

pol <- corn_yield_st \%>\%
     st_transform(2163) \%>\%
  rename(target = data) \%>\%
  nest(data = c(geometry)) \%>\%
  rename(pol = data) \%>\%
  filter(county == "polk") ;pol
  # humboldt for none; allamakee for one; boone for two
pol <- pol$pol[[1]]; pol
data_stN <- water_Temperature_st \%>\% st_transform(2163)
parm_nm = "watertem"
plot(st_geometry(pol))
plot(st_geometry(data_stN), add = TRUE)
pol_cropping(data_stN = water_Temperature_st, pol = pol, parm_nm = "watertem")
}
