% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictor_drop_null.R
\name{predictor_drop_null}
\alias{predictor_drop_null}
\title{drop rows where the predictor is NULL for stN class object}
\usage{
predictor_drop_null(data_stN)
}
\description{
drop rows where the predictor is NULL for stN class object
}
