% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescaling_summarise.R
\name{rescaling_summarise}
\alias{rescaling_summarise}
\title{rescaling the temporal resolution by method}
\usage{
rescaling_summarise(data, rescallingMethod)
}
\arguments{
\item{data}{a list of numeric data}

\item{rescallingMethod}{a string elements have to be those destinaGroup generic methods,
such as min, max, mean, etc.(see: ??dplyr::summarise)}
}
\description{
rescaling the temporal resolution by method
}
\examples{
data <- c(NA,round(runif(n = 20, min = 0, max = 100)))
rescallingMethod <- "max"
rescaling_summarise(data= data, rescallingMethod = rescallingMethod)
}
