% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RoundCoords.R
\name{RoundCoords}
\alias{RoundCoords}
\title{Function to round teh coordinates, used to de-identify data. Coordinates must be in UTM}
\usage{
RoundCoords(Coords, fct)
}
\arguments{
\item{Coords}{The coordinates that want to round.}

\item{fct}{A factor at which the coordinates will be rounded}
}
\description{
Function to round teh coordinates, used to de-identify data. Coordinates must be in UTM
}
