% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShiftCoords.R
\name{ShiftCoords}
\alias{ShiftCoords}
\title{Function to Shift a set of coordinates, used to de-identify data}
\usage{
ShiftCoords(x)
}
\arguments{
\item{x}{The set of coordinates to be used, must be in a data.frame format}
}
\description{
Function to Shift a set of coordinates, used to de-identify data
}
