% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatio_fuse.R
\name{spatio_fuse}
\alias{spatio_fuse}
\title{sort sites by geometry polygon}
\usage{
spatio_fuse(target_stN = NULL, data_stN = NULL, parm_nm = "parm_xxx",
  crs = 4326)
}
\value{
a dataframe
}
\description{
matching the site from data_stN to the by polygon in target_stN
}
\examples{
data(weather_tmin_sf)
data(corn_yield_sf)
corn_yield_st <- corn_yield_sf
weather_tmin_st <- weather_tmin_sf
tictoc::tic()
target_data_t <- spatio_fuse(target_stN = corn_yield_st,
                             data_stN = weather_tmin_st,
                             parm_nm = "tmin",
                             crs = 2163); target_data_t
tictoc::toc()
}
