% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_overlaps_with_threshold.R
\name{st_overlaps_with_threshold}
\alias{st_overlaps_with_threshold}
\title{Function Overlap based on a treshold}
\usage{
st_overlaps_with_threshold(x, y, threshold)
}
\arguments{
\item{x}{spatial object}

\item{y}{spatial object to overlap with}

\item{treshold}{A treshold for the overlap}
}
\description{
Function Overlap based on a treshold
}
