% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stN_drop_null.R
\name{stN_drop_null}
\alias{stN_drop_null}
\title{drop tows where the predictor is NULL for stN class object}
\usage{
stN_drop_null(data_stN)
}
\description{
drop tows where the predictor is NULL for stN class object
}
