% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_rescaling.R
\name{temporal_rescaling}
\alias{temporal_rescaling}
\title{rescaling the temporal resolution by method}
\usage{
temporal_rescaling(data, rescallingMethod)
}
\arguments{
\item{data}{a list of numeric data}

\item{rescallingMethod}{a string elements have to be those destinaGroup generic methods,
such as min, max, mean, etc.(see: ??dplyr::summarise)}
}
\description{
rescaling the temporal resolution by method
}
\examples{
data <- c(NA,round(runif(n = 20, min = 0, max = 100)))
rescallingMethod <- "max"
temporal_rescaling(data= data, rescallingMethod = rescallingMethod)
}
