% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cropping_cbind.R
\name{cropping_cbind}
\alias{cropping_cbind}
\title{assign site to pol by st_intersection()}
\usage{
cropping_cbind(data_stN = NULL, pol = NULL, parm_nm = "parm_nm")
}
\description{
assign site to pol by st_intersection()
}
\examples{
data(weather_tmin_sf)
data(corn_yield_sf)
corn_yield_st <- corn_yield_sf
weather_tmin_st <- weather_tmin_sf

pol <- corn_yield_st \%>\%
  rename(target = data) \%>\%
  nest(data = c(geometry)) \%>\%
  rename(pol = data) \%>\%
  ilter(county == "boone");pol # humboldt for none; allamakee for one; boone for two
pol <- pol$pol[[1]];pol
parm_nm = "tmin"
plot(st_geometry(pol))
plot(st_geometry(data_stN), add = TRUE)
cropping_cbind(data_stN = weather_tmin_st, pol = pol, parm_nm = "tmin")
}
