% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatio_fuse.R
\name{spatio_fuse}
\alias{spatio_fuse}
\title{sort sites by geometry polygon}
\usage{
spatio_fuse(target_stN = NULL, data_stN = NULL, parm_nm = "parm_xxx",
  crs = 4326)
}
\value{
a dataframe
}
\description{
matching the site from data_stN to the by polygon in target_stN
}
\examples{
data(weather_tmax_sf)
data(corn_yield_sf)
tictoc::tic()
target_data_sf <- spatio_fuse(target_stN = corn_yield_sf,
                               data_stN = weather_tmax_sf,
                               parm_nm = "tmax",
                               crs = 2163); target_data_sf
tictoc::toc()
}
