% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempo_fuse.R
\name{tempo_fuse}
\alias{tempo_fuse}
\title{combine target data and site within polygon}
\usage{
tempo_fuse(target_data = NULL, date_col = c("Year", "Date"),
  scaling = c("Year", "Day"), aggMethod = c("mean", "max"))
}
\description{
combine target data and site within polygon
}
\examples{
data(weather_tmax_sf);
data(corn_yield_sf);
target_data <- spatio_fuse(target_stN = corn_yield_sf,
                            data_stN = weather_tmax_sf,
                            parm_nm = "tmax",
                            crs = 4326); target_data
target_data <- tempo_fuse(target_data = target_data,
                          date_col = c("Year", "Date"),
                          scaling = c("Year","Month"),
                          aggMethod = c("mean","max")); target_data
}
